% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_intervals}
\alias{assert_intervals}
\title{Assertion for vector describing intervals}
\usage{
assert_intervals(x, y)
}
\arguments{
\item{x}{what to check}

\item{y}{required length of y}
}
\value{
Raises an error if
\code{x} is not a vector of length y or
not all elements of \code{x} are non-negative numbers
not all elements of \code{x} are unique and sorted or
its first element is not 0  or
}
\description{
Assertion for vector describing intervals
}
\examples{
assert_intervals(c(0,5,7),3)
}
