% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_positive_number}
\alias{assert_positive_number}
\title{Assertion for Positive Number}
\usage{
assert_positive_number(x, zero_ok = FALSE)
}
\arguments{
\item{x}{what to check.}

\item{zero_ok}{(\code{flag})\cr whether \code{x} can be zero or not.}
}
\value{
Raises an error if \code{x} is not a single positive (or non-negative) number.
}
\description{
Assertion for Positive Number
}
\examples{
assert_positive_number(3.2)
assert_positive_number(0, zero_ok = TRUE)
}
