% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corPFSOS.R
\name{corTrans}
\alias{corTrans}
\title{Correlation of PFS and OS event times for Different Transition Models}
\usage{
corTrans(transition)
}
\arguments{
\item{transition}{(\code{TransitionParameters})\cr
see \code{\link[=exponential_transition]{exponential_transition()}}, \code{\link[=weibull_transition]{weibull_transition()}} or \code{\link[=piecewise_exponential]{piecewise_exponential()}} for details.}
}
\value{
The correlation of PFS and OS.
}
\description{
Correlation of PFS and OS event times for Different Transition Models
}
\examples{
transition <- exponential_transition(h01 = 1.2, h02 = 1.5, h12 = 1.6)
corTrans(transition)
}
