% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitionParameters.R
\name{cox_transition}
\alias{cox_transition}
\title{Transition Hazards for Cox Event Times}
\usage{
cox_transition(h01 = 1, h02 = 1, h12 = 1)
}
\arguments{
\item{h01}{(positive \code{number})\cr baseline transition hazard for 0 to 1 transition.}

\item{h02}{(positive \code{number})\cr baseline transition hazard for 0 to 2 transition.}

\item{h12}{(positive \code{number})\cr baseline transition hazard for 1 to 2 transition.}
}
\value{
List with elements \code{hazards}, \code{intervals}, \code{weibull_rates} and \code{family}
(cox).
}
\description{
This creates a list with class \code{TransitionParameters} containing
baseline hazards, time intervals and Weibull rates for Cox event times
in an illness-death model.
}
\examples{
cox_transition(1, 1.6, 0.3)
}
