% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazardFunctions.R
\name{ExpHazOS}
\alias{ExpHazOS}
\title{OS Hazard Function from Constant Transition Hazards}
\usage{
ExpHazOS(t, h01, h02, h12)
}
\arguments{
\item{t}{(\code{numeric})\cr study time-points.}

\item{h01}{(positive \code{number})\cr transition hazard for 0 to 1 transition.}

\item{h02}{(positive \code{number})\cr transition hazard for 0 to 2 transition.}

\item{h12}{(positive \code{number})\cr transition hazard for 1 to 2 transition.}
}
\value{
This returns the value of the OS hazard function at time t.
}
\description{
OS Hazard Function from Constant Transition Hazards
}
\examples{
ExpHazOS(c(1:5), 0.2, 1.1, 0.8)
}
