% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitionParameters.R
\name{exponential_transition}
\alias{exponential_transition}
\title{Class of transition hazards for exponentially distributed survival times}
\usage{
exponential_transition(h01, h02, h12)
}
\arguments{
\item{h01}{transition hazard for 0 to 1 transition}

\item{h02}{transition hazard for 0 to 2 transition}

\item{h12}{transition hazard for 1 to 2 transition}
}
\value{
list of class of TransitionParameters for exponentially distributed survival times
}
\description{
Class of transition hazards for exponentially distributed survival times
}
\examples{
 exponential_transition(1, 1.6, 0.3)
}
