% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivalFunctions.R
\name{ExpSurvPFS}
\alias{ExpSurvPFS}
\title{PFS survival function from constant transition hazards.}
\usage{
ExpSurvPFS(t, h01, h02)
}
\arguments{
\item{t}{(positive \code{number})\cr  study time-point.}

\item{h01}{(positive \code{number})\cr transition hazard for 0 to 1 transition.}

\item{h02}{(positive \code{number})\cr transition hazard for 0 to 2 transition.}
}
\value{
This returns the value of PFS survival function at time t.
}
\description{
PFS survival function from constant transition hazards.
}
\examples{
ExpSurvPFS(5, 0.2, 0.4)
}
