% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateParams.R
\name{getInit.ExponentialTransition}
\alias{getInit.ExponentialTransition}
\title{Retrieve Initial Parameters for Exponential Transition Model}
\usage{
\method{getInit}{ExponentialTransition}(transition)
}
\arguments{
\item{transition}{(\code{ExponentialTransition})\cr containing the initial parameters.
See \code{\link[=exponential_transition]{exponential_transition()}} for details.}
}
\value{
Returns a vector (\code{numeric}) of initial parameters for the exponential transition.
}
\description{
Retrieve Initial Parameters for Exponential Transition Model
}
\details{
Extracts initial parameter values specifically for an exponential transition model.
}
\examples{
transition <- exponential_transition(h01 = 1.2, h02 = 1.5, h12 = 1.6)
getInit(transition)
}
