% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateParams.R
\name{getInit.WeibullTransition}
\alias{getInit.WeibullTransition}
\title{Retrieve Initial Parameters for Weibull Transition Model}
\usage{
\method{getInit}{WeibullTransition}(transition)
}
\arguments{
\item{transition}{(\code{WeibullTransition})\cr containing the initial parameters.
See \code{\link[=weibull_transition]{weibull_transition()}} for details.}
}
\value{
Returns a vector (\code{numeric}) of initial parameters for the Weibull transition.
}
\description{
Retrieve Initial Parameters for Weibull Transition Model
}
\details{
Extracts initial parameter values specifically for a Weibull transition model.
}
\examples{
transition <- weibull_transition(h01 = 1.2, h02 = 1.5, h12 = 1.6, p01 = 2, p02 = 2.5, p12 = 3)
getInit(transition)
}
