% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateParams.R
\name{getInitial}
\alias{getInitial}
\title{Retrieve Initial Parameter Vectors for Likelihood Maximization}
\usage{
getInitial(transition)
}
\arguments{
\item{transition}{(\code{ExponentialTransition} or \code{WeibullTransition})\cr containing the initial parameters.
See \code{\link[=exponential_transition]{exponential_transition()}} or \code{\link[=weibull_transition]{weibull_transition()}} for details.}
}
\value{
Returns a vector (\code{numeric}) of initial parameters for likelihood maximization.
}
\description{
Retrieve Initial Parameter Vectors for Likelihood Maximization
}
\details{
This function dispatches to either \code{getInitial.ExponentialTransition} or \code{getInitial.WeibullTransition}
based on the \code{transition} object class.
}
\examples{
transition <- exponential_transition(h01 = 1.2, h02 = 1.5, h12 = 1.6)
getInitial(transition)
}
