% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventTracking.R
\name{getNumberEvents}
\alias{getNumberEvents}
\title{Helper Function for \code{trackEventsPerTrial}}
\usage{
getNumberEvents(event, time, t)
}
\arguments{
\item{event}{(\code{numeric})\cr event indicator.}

\item{time}{(\code{numeric}) \cr event times.}

\item{t}{(\code{numeric})\cr  study time-point.}
}
\value{
This function returns the number of events occurred until time t.
}
\description{
Helper Function for \code{trackEventsPerTrial}
}
\examples{
event <- c(0, 1, 1, 1, 0)
time <- c(3, 3.4, 5, 6, 5.5)
getNumberEvents(event = event, time = time, t = 5)
}
