% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewiseDistribution.R
\name{getPCWDistr}
\alias{getPCWDistr}
\title{Piecewise exponentially distributed event times}
\usage{
getPCWDistr(U, haz, pw, t_0)
}
\arguments{
\item{U}{(\verb{numeric vector})\cr uniformly distributed random variables.}

\item{haz}{(\verb{numeric vector})\cr piecewise constant hazard.}

\item{pw}{(\verb{numeric vector})\cr time intervals for the piecewise constant hazard.}

\item{t_0}{(\verb{numeric vector})\cr the starting times.}
}
\value{
This returns a vector with event times.
}
\description{
This returns event times with a distribution resulting from piece-wise constant hazards
using the inversion method.
}
\examples{
getPCWDistr(runif(3), c(1.1, 0.5, 0.4), c(0, 7, 10), c(0, 1, 4.2))
}
