% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewise_hazards.R
\name{getPCWHazard}
\alias{getPCWHazard}
\title{Piecewise Constant Hazard Values}
\usage{
getPCWHazard(haz, pw, x)
}
\arguments{
\item{haz}{(\code{numeric})\cr input hazards.}

\item{pw}{bla.}

\item{x}{bla.}
}
\value{
Bli.
}
\description{
This calculates piecewise constant hazard values ...
}
\examples{
getPCWHazard(1:5, 2:3, 2)
}
