% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateParams.R
\name{getResults.ExponentialTransition}
\alias{getResults.ExponentialTransition}
\title{Format Results of Parameter Estimation for Exponential Transition}
\usage{
\method{getResults}{ExponentialTransition}(transition, res)
}
\arguments{
\item{transition}{(\code{ExponentialTransition})\cr
See \code{\link[=exponential_transition]{exponential_transition()}} for details.}

\item{res}{(\code{numeric} vector)\cr vector of parameter estimates from the likelihood maximization procedure.}
}
\value{
Returns an \code{ExponentialTransition} object with updated parameter estimates.
}
\description{
Format Results of Parameter Estimation for Exponential Transition
}
\details{
Constructs an \code{ExponentialTransition} object with parameters estimated from likelihood maximization.
}
\examples{
results <- c(1.2, 1.5, 1.6)
getResults(exponential_transition(), results)
}
