% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateParams.R
\name{getResults}
\alias{getResults}
\title{Format Results of Parameter Estimation for Different Transition Models}
\usage{
getResults(transition, res)
}
\arguments{
\item{transition}{(\code{ExponentialTransition} or \code{WeibullTransition})\cr
See \code{\link[=exponential_transition]{exponential_transition()}} or \code{\link[=weibull_transition]{weibull_transition()}} for details.}

\item{res}{(\code{numeric} vector)\cr vector of parameter estimates from the likelihood maximization procedure.}
}
\value{
Returns a \code{TransitionParameters} object with parameter estimates.
}
\description{
Format Results of Parameter Estimation for Different Transition Models
}
\details{
This function dispatches to either \code{getResults.ExponentialTransition} or \code{getResults.WeibullTransition}
based on the \code{transition} object class, using the results from likelihood maximization.
}
\examples{
transition <- exponential_transition(h01 = 1.2, h02 = 1.5, h12 = 1.6)
results <- c(1.2, 1.5, 1.6)
getResults(transition, results)
}
