% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSimulatedData.R
\name{getSimulatedData}
\alias{getSimulatedData}
\title{Simulated data set from an illness-death model.}
\usage{
getSimulatedData(
  N,
  transition = exponential_transition(h01 = 1, h02 = 1, h12 = 1),
  dropout = list(rate = 0, time = 12),
  accrual = list(param = "time", value = 0)
)
}
\arguments{
\item{N}{(\code{int})\cr number of patients.}

\item{transition}{(\code{TransitionParameters})\cr transition parameters comprising
\code{hazards}, corresponding \code{intervals} and \code{weibull_rates}, see \code{exponential_transition }, \code{piecewise_exponential}
and \code{weibull_transition} for details.}

\item{dropout}{(\code{list})\cr specifies drop-out probability.
Random censoring times are generated using exponential distribution. \code{dropout$rate} specifies
the drop-out probability
per \code{dropout$time} time units. If \code{dropout$rate}is equal 0, no censoring is applied.}

\item{accrual}{(\code{list})\cr specifies accrual intensity. See \code{addStaggeredEntry} for details.}
}
\value{
This returns a data frame with one row per transition per individual.
}
\description{
This function creates a single simulated data set for a single treatment arm. It simulates data
from an illness-death model with one row per transition and subject.
}
\examples{
getSimulatedData(10,
  transition = exponential_transition(h01 = 1, h02 = 1.5, h12 = 1),
  dropout = list(rate = 0.3, time = 1),
  accrual = list(param = "time", value = 5)
)
}
