% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateParams.R
\name{getTarget}
\alias{getTarget}
\title{Generate the Target Function for Optimization}
\usage{
getTarget(params, data, transition)
}
\arguments{
\item{params}{(\code{numeric} vector)\cr
Vector of parameters to be optimized over.}

\item{data}{(\code{data.frame})\cr
Data frame in the format created by \code{\link[=prepareData]{prepareData()}}.}

\item{transition}{(\code{transitionParameters} object)\cr
Object specifying the distribution family. See \code{\link[=exponential_transition]{exponential_transition()}} or \code{\link[=weibull_transition]{weibull_transition()}} for details.}
}
\value{
Returns a function that calculates the negative log-likelihood (\code{numeric}) for the given parameters.
}
\description{
Generate the Target Function for Optimization
}
\details{
This function creates a target function for optimization, which computes the negative log-likelihood for given
parameters, data, and transition model type.
}
\examples{
params <- c(1.2, 1.5, 1.6) # For ExponentialTransition
data <- prepareData(your_dataset)
transition <- exponential_transition()
getTarget(params, data, transition)
}
