% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateParams.R
\name{haz.ExponentialTransition}
\alias{haz.ExponentialTransition}
\title{Hazard Function for Exponential Transition Model}
\usage{
\method{haz}{ExponentialTransition}(transition, t, trans)
}
\arguments{
\item{transition}{(\code{ExponentialTransition})\cr
See \code{\link[=exponential_transition]{exponential_transition()}} for details.}

\item{t}{(\code{numeric})\cr time at which hazard is to be computed.}

\item{trans}{(\code{integer})\cr index specifying the transition type.}
}
\value{
Returns the hazard rate (\code{numeric}) for the exponential transition at the specified time.
}
\description{
Hazard Function for Exponential Transition Model
}
\details{
Computes the hazard function specifically for an exponential transition model using provided parameters.
}
\examples{
transition <- exponential_transition(h01 = 1.2, h02 = 1.5, h12 = 1.6)
haz.ExponentialTransition(transition, 0.4, 2)
}
