% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateParams.R
\name{haz}
\alias{haz}
\title{Hazard Function for Different Transition Models}
\usage{
haz(transition, t, trans)
}
\arguments{
\item{transition}{(\code{ExponentialTransition} or \code{WeibullTransition})\cr.
See \code{\link[=exponential_transition]{exponential_transition()}} or \code{\link[=weibull_transition]{weibull_transition()}} for details.}

\item{t}{(\code{numeric})\cr time at which hazard is to be computed.}

\item{trans}{(\code{integer})\cr index specifying the transition type.}
}
\value{
Returns the hazard rate (\code{numeric}) for the specified transition and time.
}
\description{
Hazard Function for Different Transition Models
}
\details{
This function dispatches to either \code{haz.ExponentialTransition} or \code{haz.WeibullTransition}
based on the \code{transition} object class.
}
\examples{
transition <- exponential_transition(h01 = 1.2, h02 = 1.5, h12 = 1.6)
haz(transition, 0.4, 2)
}
