% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateParams.R
\name{haz.WeibullTransition}
\alias{haz.WeibullTransition}
\title{Hazard Function for Weibull Transition Model}
\usage{
\method{haz}{WeibullTransition}(transition, t, trans)
}
\arguments{
\item{transition}{(\code{WeibullTransition})\cr
See \code{\link[=weibull_transition]{weibull_transition()}} for details.}

\item{t}{(\code{numeric})\cr time at which hazard is to be computed.}

\item{trans}{(\code{integer})\cr index specifying the transition type.}
}
\value{
Returns the hazard rate (\code{numeric}) for the Weibull transition at the specified time.
}
\description{
Hazard Function for Weibull Transition Model
}
\details{
Computes the hazard function specifically for a Weibull transition model using provided parameters.
}
\examples{
transition <- weibull_transition(h01 = 1.2, h02 = 1.5, h12 = 1.6, p01 = 2, p02 = 2.5, p12 = 3)
haz(transition, 0.4, 2)
}
