% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitionParameters.R
\name{make_betas}
\alias{make_betas}
\title{Parameter vectors for Cox model}
\usage{
make_betas(b01, b02, b12)
}
\arguments{
\item{b01}{(numeric vector)\cr betas for 0 to 1 transition.}

\item{b02}{(numeric vector)\cr betas for 0 to 2 transition.}

\item{b12}{(numeric vector)\cr betas for 1 to 1 transition.}
}
\value{
List with three beta vectors as elements.
}
\description{
This creates a list containing the parameter vectors for the Cox model
for the three transitions in an illness-death model.
}
