% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corPFSOS.R
\name{PFSOSInteg}
\alias{PFSOSInteg}
\title{Helper Function for \code{survPFSOS()}}
\usage{
PFSOSInteg(u, t, transition)
}
\arguments{
\item{u}{(\code{numeric})\cr variable of integration.}

\item{t}{(\code{numeric})\cr time at which the value of the PFS*OS survival function is to be computed.}

\item{transition}{(\code{TransitionParameters})\cr
see \code{\link[=exponential_transition]{exponential_transition()}}, \code{\link[=weibull_transition]{weibull_transition()}} or \code{\link[=piecewise_exponential]{piecewise_exponential()}} for details.}
}
\value{
Numeric result of the integrand used to calculate the PFS*OS survival function.
}
\description{
Helper Function for \code{survPFSOS()}
}
\note{
Not all vectors \code{u} and \code{t} work here due to assertions in \code{\link[=log_p11]{log_p11()}}.
}
\keyword{internal}
