% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitionParameters.R
\name{piecewise_exponential}
\alias{piecewise_exponential}
\title{Class of transition hazards for piecewise exponentially distributed survival times}
\usage{
piecewise_exponential(h01, h02, h12, pw01, pw02, pw12)
}
\arguments{
\item{h01}{vector of constant transition hazards for 0 to 1 transition}

\item{h02}{vector of constant transition hazards for 0 to 2 transition}

\item{h12}{vector of constant transition hazards for 1 to 2 transition}

\item{pw01}{vector of time intervals for the piecewise constant hazards \code{h01}}

\item{pw02}{vector of time intervals for the piecewise constant hazards \code{h02}}

\item{pw12}{vector of time intervals for the piecewise constant hazards \code{h12}}
}
\value{
list of class of TransitionParameters for piecewise exponentially distributed survival times
}
\description{
Class of transition hazards for piecewise exponentially distributed survival times
}
\examples{
piecewise_exponential(h01 = c(1, 1, 1),h02 = c(1.5, 0.5, 1),h12 = c(1, 1, 1),
pw01 = c(0, 3, 8),pw02 = c(0, 6, 7),pw12 = c(0, 8, 9))
}
