% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivalFunctions.R
\name{pwA}
\alias{pwA}
\title{Cumulative hazard for piecewise constant hazards}
\usage{
pwA(t, haz, pw)
}
\arguments{
\item{t}{(positive \code{number})\cr  study time-point.}

\item{haz}{(\verb{numeric vector})\cr constant transition hazards.}

\item{pw}{(\verb{numeric vector})\cr time intervals for the piecewise constant hazards.}
}
\value{
This returns the value of cumulative hazard at time t.
}
\description{
Cumulative hazard for piecewise constant hazards
}
\examples{
pwA(5, c(0.5, 0.9), c(0, 4))
}
