% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClinicalTrials.R
\name{runTrial}
\alias{runTrial}
\title{Helper Function for Adding Progress Bar to Trial Simulation}
\usage{
runTrial(x, pb, ...)
}
\arguments{
\item{x}{(\code{int})\cr iteration index within lapply.}

\item{...}{parameters transferred to \code{\link[=getOneClinicalTrial]{getOneClinicalTrial()}}, see \code{\link[=getOneClinicalTrial]{getOneClinicalTrial()}} for details.}
}
\value{
This returns the same as \code{\link[=getOneClinicalTrial]{getOneClinicalTrial()}}, but updates the progress bar.
}
\description{
Helper Function for Adding Progress Bar to Trial Simulation
}
\keyword{internal}
