% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{SimulationEngineMSM-package}
\alias{SimulationEngineMSM}
\alias{SimulationEngineMSM-package}
\title{\code{SimulationEngineMSM} Package}
\description{
\code{SimulationEngineMSM} does ...
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/insightsengineering/SimulationEngineMSM/}
  \item Report bugs at \url{https://github.com/insightsengineering/SimulationEngineMSM/issues}
}

}
\author{
\strong{Maintainer}: Alexandra Niessl \email{niessl.alexandra@gene.com}

}
