% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corPFSOS.R
\name{survPFS}
\alias{survPFS}
\title{PFS Survival Function for Different Transition Models}
\usage{
survPFS(transition, t)
}
\arguments{
\item{transition}{(\code{TransitionParameters})\cr
see \code{\link[=exponential_transition]{exponential_transition()}}, \code{\link[=weibull_transition]{weibull_transition()}} or \code{\link[=piecewise_exponential]{piecewise_exponential()}} for details.}

\item{t}{(\code{numeric})\cr time at which the value of the PFS survival function is to be computed.}
}
\value{
The value of the survival function for the specified transition and time.
}
\description{
PFS Survival Function for Different Transition Models
}
\examples{
transition <- exponential_transition(h01 = 1.2, h02 = 1.5, h12 = 1.6)
survPFS(transition, 0.4)
}
