% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateParams.R
\name{survTrans.ExponentialTransition}
\alias{survTrans.ExponentialTransition}
\title{Survival Function for Exponential Transition Model}
\usage{
\method{survTrans}{ExponentialTransition}(transition, t, trans)
}
\arguments{
\item{transition}{(\code{ExponentialTransition})\cr
See \code{\link[=exponential_transition]{exponential_transition()}} for details.}

\item{t}{(\code{numeric})\cr time at which survival probability is to be computed.}

\item{trans}{(\code{integer})\cr index specifying the transition type.}
}
\value{
Returns the survival probability (\code{numeric}) for the exponential transition at the specified time.
}
\description{
Survival Function for Exponential Transition Model
}
\details{
Computes the survival function specifically for an exponential transition model using provided parameters.
}
\examples{
transition <- exponential_transition(h01 = 1.2, h02 = 1.5, h12 = 1.6)
survTrans.ExponentialTransition(transition, 0.4, 2)
}
