% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitionParameters.R
\name{weibull_transition}
\alias{weibull_transition}
\title{todo: similar as above
Class of transition parameters describing Weibull distributed survival times
in an illness-death model}
\usage{
weibull_transition(h01, h02, h12, p01, p02, p12)
}
\arguments{
\item{h01}{transition hazard for 0 to 1 transition}

\item{h02}{transition hazard for 0 to 2 transition}

\item{h12}{transition hazard for 1 to 2 transition}

\item{p01}{rate parameter of Weibull distribution for \code{h01}}

\item{p02}{rate parameter of Weibull distribution for \code{h02}}

\item{p12}{rate parameter of Weibull distribution for \code{h12}}
}
\value{
list of transition parameters for Weibull distributed survival times in
an illness-death model
}
\description{
todo: similar as above
Class of transition parameters describing Weibull distributed survival times
in an illness-death model
}
\examples{
weibull_transition(h01 = 1, h02 = 1.3, h12 = 0.5, p01 = 1.2, p02 = 1.3, p12 = 0.5)
}
