% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_intervals}
\alias{assert_intervals}
\title{Assertion for vector describing intervals}
\usage{
assert_intervals(x, y)
}
\arguments{
\item{x}{what to check.}

\item{y}{(\code{count})\cr required length of \code{y}.}
}
\value{
Raises an error if \code{x} is not an intervals vector starting with 0.
}
\description{
We define an intervals vector to always start with 0, and contain
unique ordered time points.
}
\examples{
assert_intervals(c(0, 5, 7), 3)
}
