% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivalFunctions.R
\name{ExpSurvPFS}
\alias{ExpSurvPFS}
\title{PFS Survival Function from Constant Transition Hazards}
\usage{
ExpSurvPFS(t, h01, h02)
}
\arguments{
\item{t}{(\code{numeric})\cr  study time-points.}

\item{h01}{(positive \code{number})\cr transition hazard for 0 to 1 transition.}

\item{h02}{(positive \code{number})\cr transition hazard for 0 to 2 transition.}
}
\value{
This returns the value of PFS survival function at time t.
}
\description{
PFS Survival Function from Constant Transition Hazards
}
\examples{
ExpSurvPFS(c(1:5), 0.2, 0.4)
}
