% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewiseHazards.R
\name{getPCWHazard}
\alias{getPCWHazard}
\title{Piecewise Constant Hazard Values}
\usage{
getPCWHazard(haz, pw, x)
}
\arguments{
\item{haz}{(\code{numeric})\cr piecewise constant input hazard.}

\item{pw}{(\code{numeric})\cr time intervals for the piecewise constant hazard.}

\item{x}{(\code{numeric})\cr time-points.}
}
\value{
Hazard values at input time-points.
}
\description{
This returns piecewise constant hazard values at specified time points.
}
\examples{
getPCWHazard(c(1, 1.2, 1.4), c(0, 2, 3), c(1, 4, 6))
}
