% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateParams.R
\name{getResults.WeibullTransition}
\alias{getResults.WeibullTransition}
\title{Format Results of Parameter Estimation for Weibull Transition}
\usage{
\method{getResults}{WeibullTransition}(transition, res)
}
\arguments{
\item{transition}{(\code{WeibullTransition})\cr
See \code{\link[=weibull_transition]{weibull_transition()}} for details.}

\item{res}{(\code{numeric} vector)\cr vector of parameter estimates from the likelihood maximization procedure.}
}
\value{
Returns a \code{WeibullTransition} object with updated parameter estimates.
}
\description{
Format Results of Parameter Estimation for Weibull Transition
}
\details{
Constructs a \code{WeibullTransition} object with parameters estimated from likelihood maximization.
}
\examples{
results <- c(1.2, 1.5, 1.6, 2, 2.5, 1)
getResults.WeibullTransition(weibull_transition(), results)
}
