% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewiseHazards.R
\name{getSumPCW}
\alias{getSumPCW}
\title{Sum of Two Piecewise Constant Hazards}
\usage{
getSumPCW(haz1, haz2, pw1, pw2)
}
\arguments{
\item{haz1}{(\code{numeric})\cr first summand (piecewise constant hazard).}

\item{haz2}{(\code{numeric})\cr second summand (piecewise constant hazard).}

\item{pw1}{(\code{numeric})\cr time intervals of first summand.}

\item{pw2}{(\code{numeric})\cr time intervals of second summand.}
}
\value{
List with elements \code{hazards} and \code{intervals} for the sum of two piecewise constant hazards.
}
\description{
This returns the sum of two piecewise constant hazards per interval.
}
\examples{
getSumPCW(c(1.2, 0.3, 0.6), c(1.2, 0.7, 1), c(0, 8, 9), c(0, 1, 4))
}
