% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivalFunctions.R
\name{integrateVector}
\alias{integrateVector}
\title{Helper function for efficient integration.}
\usage{
integrateVector(integrand, upperVector, ...)
}
\arguments{
\item{integrand}{an \code{R} function to be integrated.}

\item{upperVector}{(\code{numeric})\cr upper limits of integration.}

\item{...}{additional arguments to be passed to \code{integrand}.}
}
\value{
This function returns for each upper limit the estimates of the integral.
}
\description{
Helper function for efficient integration.
}
\examples{
integrand <- function(x) x^2
upperVector <- c(0, 1, 0.4, 2, 5, 2, 0.3, 0.4, 1)
integrateVector(integrand, upperVector = upperVector)
}
