% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivalFunctions.R
\name{integrateVector}
\alias{integrateVector}
\title{Helper for Efficient Integration}
\usage{
integrateVector(integrand, upper, ...)
}
\arguments{
\item{integrand}{(\code{function})\cr  to be integrated.}

\item{upper}{(\code{numeric})\cr upper limits of integration.}

\item{...}{additional arguments to be passed to \code{integrand}.}
}
\value{
This function returns for each upper limit the estimates of the integral.
}
\description{
Helper for Efficient Integration
}
\keyword{internal}
