% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corPFSOS.R
\name{p11Integ}
\alias{p11Integ}
\title{Helper Function for \code{log_p11()}}
\usage{
p11Integ(x, transition)
}
\arguments{
\item{x}{(\code{numeric})\cr variable of integration.}

\item{transition}{(\code{TransitionParameters})\cr
see \code{\link[=exponential_transition]{exponential_transition()}}, \code{\link[=weibull_transition]{weibull_transition()}} or \code{\link[=piecewise_exponential]{piecewise_exponential()}} for details.}
}
\value{
Hazard rate at the specified time for the transition from progression to death.
}
\description{
Helper Function for \code{log_p11()}
}
\keyword{internal}
