% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivalFunctions.R
\name{pwA}
\alias{pwA}
\title{Cumulative Hazard for Piecewise Constant Hazards}
\usage{
pwA(t, haz, pw)
}
\arguments{
\item{t}{(\code{numeric})\cr  study time-points.}

\item{haz}{(\verb{numeric vector})\cr constant transition hazards.}

\item{pw}{(\verb{numeric vector})\cr time intervals for the piecewise constant hazards.}
}
\value{
This returns the value of cumulative hazard at time t.
}
\description{
Cumulative Hazard for Piecewise Constant Hazards
}
\examples{
pwA(1:5, c(0.5, 0.9), c(0, 4))
}
