% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{simIDM-package}
\alias{simIDM}
\alias{simIDM-package}
\title{\code{simIDM} Package}
\description{
\code{simIDM} simulates a survival multistate model that jointly models PFS and OS.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/insightsengineering/simIDM/}
  \item Report bugs at \url{https://github.com/insightsengineering/simIDM/issues}
}

}
\author{
\strong{Maintainer}: Alexandra Erdmann \email{alexandra.erdmann@uni-ulm.de}

Authors:
\itemize{
  \item Kaspar Rufibach \email{kaspar.rufibach@roche.com}
}

Other contributors:
\itemize{
  \item Daniel Sabanes Bove \email{daniel.sabanes_bove@roche.com} [contributor]
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
  \item University of Ulm [copyright holder, funder]
}

}
